updated slideshow with reddit articles
